requirements.txt
python-dotenv>=0.19.0
requests>=2.31.0
logfire>=0.10.0